Add README, package, maven changes
