+++ getting-started-with-api/README.md
@@ -3,3 +3,4 @@
 Scripts illustrating topics from the [Getting Started](https://help.smartling.com/hc/en-us/articles/1260804661570-Getting-Started) tutorial.
 
 * [Python](python)
+* [Java-SDK](java-sdk)
