+++ getting-started-with-api/java-sdk/pom.xml
@@ -4,7 +4,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>smartling-samples</groupId>
+  <groupId>com.smartling.samples</groupId>
   <artifactId>getting-started</artifactId>
   <version>1.0-SNAPSHOT</version>
 
@@ -13,8 +13,8 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
   </properties>
 
   <dependencies>
@@ -23,6 +23,49 @@
           <artifactId>smartling-api-sdk</artifactId>
           <version>0.19.0</version>
       </dependency>
-  </dependencies>
+      <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-api</artifactId>
+        <version>1.7.30</version>
+    </dependency>
+    <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-simple</artifactId>
+        <version>1.7.30</version>
+    </dependency>
+</dependencies>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>1.6.0</version>
+        <executions>
+          <execution>
+            <id>upload</id>
+            <configuration>
+              <mainClass>com.smartling.samples.gettingstarted.Upload</mainClass>
+              <arguments>
+                  <argument>strings.json</argument>
+                  <argument>json</argument>
+              </arguments>
+            </configuration>
+          </execution>
+          <execution>
+            <id>download</id>
+            <configuration>
+              <mainClass>com.smartling.samples.gettingstarted.DownloadPseudo</mainClass>
+              <arguments>
+                  <argument>strings.json</argument>
+                  <argument>strings_pseudo.json</argument>
+                  <argument>fr-FR</argument>
+              </arguments>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
 
 </project>
