+++ getting-started-with-api/java-sdk/src/main/java/com/smartling/samples/gettingstarted/DownloadPseudo.java
@@ -1,8 +1,9 @@
+package com.smartling.samples.gettingstarted;
+
 import com.smartling.api.files.v2.FilesApi;
 import com.smartling.api.files.v2.FilesApiFactory;
-import com.smartling.api.files.v2.pto.FileType;
-import com.smartling.api.files.v2.pto.UploadFilePTO;
-import com.smartling.api.files.v2.pto.UploadFileResponse;
+import com.smartling.api.files.v2.pto.DownloadTranslationPTO;
+import com.smartling.api.files.v2.pto.RetrievalType;
 import com.smartling.api.v2.authentication.AuthenticationApi;
 import com.smartling.api.v2.authentication.AuthenticationApiFactory;
 import com.smartling.api.v2.client.ClientFactory;
@@ -10,29 +11,25 @@ import com.smartling.api.v2.client.DefaultClientConfiguration;
 import com.smartling.api.v2.client.auth.Authenticator;
 import com.smartling.api.v2.client.auth.BearerAuthSecretFilter;
 import com.smartling.api.v2.client.exception.RestApiRuntimeException;
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.io.FileUtils;
+import java.io.InputStream;
+import java.io.File;
+import static java.nio.charset.StandardCharsets.UTF_8;
 
-import java.io.FileInputStream;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public class Upload
+public class DownloadPseudo
 {
-    private static final List<String> locales = new ArrayList<>(Arrays.asList("fr-FR"));
-
     public static void main( String[] args ) throws Exception
     {
         // Read command-line arguments
-        if (args.length != 2) {
-            System.out.println("Usage: mvn -X exec:java -Dexec.mainClass=\"Upload\" -Dexec.args=\"<fileName> <fileType>\"");
-            System.out.println("E.g.: mvn -X exec:java -Dexec.mainClass=\"Upload\" -Dexec.args=\"strings.json json\"");
-            System.out.println("Takes two command-line parameters: ");
-            System.out.println("  fileName: local name of file to be uploaded");
-            System.out.println("  fileType: file type of file to be uploaded");
+        if (args.length != 3) {
+            System.out.println("Usage: mvn -q exec:java@download");
+            System.out.println("(See pom.xml to configure parameters.)");
             System.exit(1);
         }
-        String fileName = args[0];
-        String fileType = args[1];
+        String fileUri = args[0];
+        String outputFileName = args[1];
+        String localeId = args[2];
 
         // Read authentication credentials from environment
         String userId = System.getenv("DEV_USER_IDENTIFIER");
@@ -56,19 +53,19 @@ public class Upload
             .buildApi(new BearerAuthSecretFilter(authenticator), clientConfiguration);
 
         // Build the API request
-        UploadFilePTO uploadFilePto = UploadFilePTO.builder()
-            .file(new FileInputStream(fileName))
-            .fileUri(fileName)
-            .fileType(FileType.lookup(fileType))
-            .localeIdsToAuthorize(locales)
+        DownloadTranslationPTO downloadTranslationPTO = DownloadTranslationPTO.builder()
+            .fileUri(fileUri)
+            .retrievalType(RetrievalType.PSEUDO)
             .build();
 
-        // Execute the upload
+        // Execute the download
         try {
-            UploadFileResponse uploadFileResponse = filesApi.uploadFile(projectId, uploadFilePto);
-            System.out.println(uploadFileResponse);
-        } catch (RestApiRuntimeException e) {
-            System.out.println(e.getMessage());
+            InputStream translatedFile = filesApi.downloadTranslatedFile(projectId, localeId, downloadTranslationPTO);
+            System.out.println("Downloaded file: " + outputFileName);
+            File outputFile = new File(outputFileName);
+            FileUtils.copyInputStreamToFile(translatedFile, outputFile);
+        } catch  (RestApiRuntimeException e) {
+            System.err.println(e.getMessage());
         }
     }
